requirements.txt
Flask
pymongo
dnspython
gunicorn
python-dotenv